Convert paramter-type utils to TypeScript
